Sample for uploading to S3 server
This commit currently does not allow one to upload from canvas to S3, however it does have the exact required parameters to create the PNG file from the base64 stream, and upload this to the S3 server. The canvas exposes this Base64 stream using: `signaturePad.toDataUrl()`.

More research is required in exactly how to get it from JS to the controller to have it pushed to the S3 server.